Made adjustments to tests.py and bindings.cpp, commented out xyResolution and GetS2 which crash for now because of ill defined FitTBA function in VDetector.  Added comment on py::nodelete in VDetector bindings
